refactor: extract table display functions to separate module
